% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomerMaterialCorrespondence.R
\name{CustomerMaterialCorrespondenceUI}
\alias{CustomerMaterialCorrespondenceUI}
\title{客户物料生成生成器界面}
\usage{
CustomerMaterialCorrespondenceUI(
  tabTitle = "客户物料对应表",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = CustomerMaterialCorrespondenceUI_left,
  func_right = CustomerMaterialCorrespondenceUI_right,
  func_bottom = CustomerMaterialCorrespondenceUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
客户物料生成生成器界面
}
\examples{
CustomerMaterialCorrespondenceUI()
}
