% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomerMaterialCorrespondence.R
\name{CustomerMaterialCorrespondenceUI_bottom}
\alias{CustomerMaterialCorrespondenceUI_bottom}
\title{生成区域}
\usage{
CustomerMaterialCorrespondenceUI_bottom()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
CustomerMaterialCorrespondenceUI_bottom()
}
