% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomerMaterialCorrespondence.R
\name{CustomerMaterialCorrespondenceUI_left}
\alias{CustomerMaterialCorrespondenceUI_left}
\title{请输入文件}
\usage{
CustomerMaterialCorrespondenceUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
CustomerMaterialCorrespondenceUI_left()
}
