% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestRecord.R
\name{TestRecordUI}
\alias{TestRecordUI}
\title{检测记录生成生成器界面}
\usage{
TestRecordUI(
  tabTitle = "检测记录",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = TestRecordUI_left,
  func_right = TestRecordUI_right,
  func_bottom = TestRecordUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
检测记录生成生成器界面
}
\examples{
TestRecordUI()
}
